requirements.txt
runpod~=1.7.9
praw
python-slugify
